Win32 (#1371)
* Make types.cc the first file on the compile list
* Support platforms without int128 in dtoa.h
* remove __restrict__ keyword
* Fix warning in unaryop.cc
* Throw an exception when trying to memmap on Windows

+++ c/expr/unaryop.cc
@@ -73,7 +73,7 @@ inline static T op_abs(T x) {
 
 template <typename T>
 inline static double op_exp(T x) {
-  return ISNA<T>(x) ? GETNA<double>() : exp(x);
+  return ISNA<T>(x) ? GETNA<double>() : exp(static_cast<double>(x));
 }
